remove print_progress() in graingrowth_MC.cpp

+++ graingrowth_MC.hpp
@@ -10,6 +10,6 @@
 std::string PROGRAM = "graingrowth";
 std::string MESSAGE = "Voronoi tessellation and isotropic grain growth code";
 
-typedef MMSP::grid<2,int> GRID2D;
-typedef MMSP::grid<3,int> GRID3D;
+typedef MMSP::grid<2,unsigned long> GRID2D;
+typedef MMSP::grid<3,unsigned long> GRID3D;
